Move Core Data Feed Store to it's own file
