+++ EssentialDeveloper/EssentialDeveloper/Feed Cache/CoreData/CoreDataFeedStore.swift
@@ -1,6 +1,6 @@
 import CoreData
 
-public final class CoreDataFeedStore: FeedStore {
+public final class CoreDataFeedStore {
     public static let modelName = "FeedStore"
     public static let model = NSManagedObjectModel(name: modelName, in: Bundle(for: CoreDataFeedStore.self))
 
@@ -23,49 +23,6 @@ public final class CoreDataFeedStore: FeedStore {
             url: storeURL
         )
         context = container.newBackgroundContext()
-    }
-    
-    public func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-        perform { context in
-            do {
-                try ManagedCache.findAndDestroy(in: context)
-                try context.save()
-
-                completion(.success(()))
-            }catch {
-                context.rollback()
-                completion(.failure(error))
-            }
-        }
-    }
-    
-    public func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-        perform { context in
-            do {
-                let managedCache = try ManagedCache.createNewUniqueInstance(in: context)
-                managedCache.timestamp = timestamp
-                managedCache.feed = ManagedFeedImage.images(from: feed, in: context)
-                
-                try context.save()
-                completion(.success(()))
-            } catch {
-                context.rollback()
-                completion(.failure(error))
-            }
-        }
-    }
-    
-    public func retrieve(completion: @escaping RetrievalCompletion) {
-        perform { context in
-            completion(Result {
-                try ManagedCache.find(in: context).map {
-                    CachedFeed(
-                        feed: $0.localFeed,
-                        timestamp: $0.timestamp
-                    )
-                }
-            })
-        }
     }
     
     public func perform(_ action: @escaping (NSManagedObjectContext) -> Void) {
